refactor: extend playwright test with network and coverage fixtures
