+++ template/tests/steps/fixture/playwrightBdd.fixture.ts
@@ -1,4 +1,15 @@
-import { createBdd } from 'playwright-bdd';
-import { test } from './testWithMsw.fixture';
+import { createNetworkFixture, type NetworkFixture } from '@msw/playwright';
+import { createBdd, test as base } from 'playwright-bdd';
+import { createTestWithCoverageFixture, TestWithCoverage } from '../../utils/createTestWithCoverageFixture';
+
+type Fixtures = {
+  network: NetworkFixture;
+  testWithCoverage: TestWithCoverage;
+};
+
+export const test = base.extend<Fixtures>({
+  network: createNetworkFixture(),
+  testWithCoverage: createTestWithCoverageFixture(),
+});
 
 export const { Given, When, Then } = createBdd(test);
