--- template/tests/steps/fixture/testWithMsw.fixture.ts
@@ -1,6 +1,0 @@
-import { createNetworkFixture, type NetworkFixture } from '@msw/playwright';
-import { test as base } from './testWithCoverage.fixture';
-
-export const test = base.extend<{ network: NetworkFixture }>({
-  network: createNetworkFixture(),
-});
